Issue #22156: Fix "comparison between signed and unsigned integers" compiler warnings in the Python/ subdirectory.

+++ Python/bltinmodule.c
@@ -581,7 +581,7 @@ source_as_string(PyObject *cmd, char *funcname, char *what, PyCompilerFlags *cf)
         return NULL;
     }
 
-    if (strlen(str) != size) {
+    if (strlen(str) != (size_t)size) {
         PyErr_SetString(PyExc_TypeError,
                         "source code string cannot contain null bytes");
         return NULL;
